+++ subprograms/ausgabe.py
